[core] add HDD vs UFD detection
* Initial scoring to try to differentiate UFDs from HDDs (#219)
* Also improve GetDriveLetter() and add a global for fixed vs removable
* Also fix a bug with reporting of VID:PID with multiple devices
* Also fix a warning in localization

+++ src/localization.c
@@ -199,7 +199,7 @@ static uint32_t htab_hash(char* str)
 		do {
 			// Because size is prime this guarantees to step through all available indexes
 			if (idx <= hval2) {
-				idx = htab_size + idx - hval2;
+				idx = ((uint32_t)htab_size) + idx - hval2;
 			} else {
 				idx -= hval2;
 			}
